requirements.txt
django
argon2-cffi
django-cors-headers
django_filters
django-storages
djangorestframework
djangorestframework_simplejwt
dj_database_url
psycopg2
whitenoise
boto3
PILLOW